pyproject.toml
[tool.poetry]
name = "mongodb-cdc-delta"
version = "0.1.0"
description = "Change Data Capture pipeline from MongoDB to Delta Lake using Kafka and Debezium."
authors = ["CDC Team <cdc@example.com>"]
readme = "README.md"
packages = [{include = "src"}]

[tool.poetry.dependencies]
python = "^3.11"
# Kafka and Debezium
kafka-python = "^2.0.2"
confluent-kafka = "^2.3.0"
# HashiCorp Vault
hvac = "^2.1.0"
# Logging and monitoring
structlog = "^24.1.0"
prometheus-client = "^0.19.0"
opentelemetry-api = "^1.21.0"
opentelemetry-sdk = "^1.21.0"
opentelemetry-instrumentation = "^0.42b0"
opentelemetry-exporter-jaeger = "^1.21.0"
# Error handling and retry
tenacity = "^8.2.3"
# FastAPI for REST API
fastapi = "^0.109.0"
uvicorn = {extras = ["standard"], version = "^0.27.0"}
# Scheduling
apscheduler = "^3.10.4"
# Testing data generation
faker = "^22.0.0"
# HTTP client for health checks and alerts
httpx = "^0.26.0"
slowapi = "^0.1.9"

[tool.poetry.group.dev.dependencies]
# Testing
pytest = "^7.4.3"
pytest-asyncio = "^0.23.2"
pytest-cov = "^4.1.0"
testcontainers = "^3.7.1"
# Code quality
ruff = "^0.1.9"
black = "^23.12.1"
mypy = "^1.8.0"
# Security
bandit = "^1.7.6"
safety = "^2.3.5"
# Pre-commit hooks
pre-commit = "^3.6.0"
# Performance benchmarking
locust = "^2.20.0"
# Property-based testing
hypothesis = "^6.92.0"
pytest-timeout = "^2.4.0"

[tool.poetry.group.test.dependencies]
pytest = "^7.4.3"
pytest-asyncio = "^0.23.2"
pytest-cov = "^4.1.0"
testcontainers = "^3.7.1"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"

[tool.pytest.ini_options]
testpaths = ["tests", "api/tests", "delta-writer/tests", "reconciliation/tests"]
python_files = ["test_*.py"]
python_classes = ["Test*"]
python_functions = ["test_*"]
addopts = [
    "--verbose",
    "-ra",
    "--strict-markers",
    "--strict-config",
    "--cov=api",
    "--cov=delta-writer",
    "--cov=reconciliation",
    "--cov=shared",
    "--cov-report=term-missing:skip-covered",
    "--cov-report=html",
    "--cov-report=xml",
    "--cov-fail-under=80",
    "--timeout=300",
    "--timeout-method=thread"
]
asyncio_mode = "auto"
timeout = 300
markers = [
    "unit: Unit tests",
    "integration: Integration tests",
    "e2e: End-to-end tests",
    "contract: Contract tests",
    "slow: Slow tests",
    "performance: Performance tests",
]

[tool.black]
line-length = 100
target-version = ['py311']
include = '\.pyi?$'
extend-exclude = '''
/(
  # directories
  \.eggs
  | \.git
  | \.hg
  | \.mypy_cache
  | \.tox
  | \.venv
  | build
  | dist
)/
'''

[tool.ruff]
line-length = 100
target-version = "py311"
select = [
    "E",  # pycodestyle errors
    "W",  # pycodestyle warnings
    "F",  # pyflakes
    "I",  # isort
    "C",  # flake8-comprehensions
    "B",  # flake8-bugbear
    "S",  # flake8-bandit (security)
]
ignore = [
    "E501",  # line too long (handled by black)
    "B008",  # do not perform function calls in argument defaults
    "S104",  # possible binding to all interfaces
    "S108",  # probable insecure usage of temp file/directory
]
exclude = [
    ".git",
    ".mypy_cache",
    ".pytest_cache",
    ".ruff_cache",
    ".venv",
    "__pycache__",
    "build",
    "dist",
]

[tool.ruff.per-file-ignores]
"tests/**/*.py" = [
    "S101",  # assert allowed in tests
    "S105",  # hardcoded passwords allowed in tests
    "S106",  # hardcoded passwords allowed in tests
]

[tool.mypy]
python_version = "3.11"
warn_return_any = true
warn_unused_configs = true
disallow_untyped_defs = true
disallow_incomplete_defs = true
check_untyped_defs = true
disallow_untyped_calls = true
disallow_untyped_decorators = false
no_implicit_optional = true
warn_redundant_casts = true
warn_unused_ignores = true
warn_no_return = true
warn_unreachable = true
strict_equality = true
ignore_missing_imports = true

[[tool.mypy.overrides]]
module = "tests.*"
disallow_untyped_defs = false
disallow_untyped_calls = false

[tool.coverage.run]
source = ["api", "delta-writer", "reconciliation", "shared"]
omit = [
    "*/tests/*",
    "*/test_*.py",
    "*/__pycache__/*",
    "*/migrations/*",
    "**/.venv/*",
]

[tool.coverage.report]
precision = 2
show_missing = true
skip_covered = false
exclude_lines = [
    "pragma: no cover",
    "def __repr__",
    "raise AssertionError",
    "raise NotImplementedError",
    "if __name__ == .__main__.:",
    "if TYPE_CHECKING:",
    "class .*\\bProtocol\\):",
    "@(abc\\.)?abstractmethod",
]

[tool.bandit]
# Bandit security linter configuration
# Checks for hardcoded passwords, secrets, and other security issues
targets = ["src"]
exclude_dirs = ["tests", ".venv", "build", "dist"]
skips = []
tests = [
    "B105",  # hardcoded_password_string
    "B106",  # hardcoded_password_funcarg
    "B107",  # hardcoded_password_default
    "B108",  # hardcoded_temp_file
    "B110",  # try_except_pass
    "B201",  # flask_debug_true
    "B301",  # pickle usage
    "B302",  # marshal usage
    "B303",  # md5 or sha1 usage
    "B304",  # ciphers usage
    "B305",  # cipher_modes usage
    "B306",  # mktemp_q usage
    "B307",  # eval usage
    "B308",  # mark_safe usage
    "B309",  # httpsconnection usage
    "B310",  # urllib_urlopen usage
    "B311",  # random usage
    "B312",  # telnetlib usage
    "B313",  # xml_bad_cElementTree
    "B314",  # xml_bad_ElementTree
    "B315",  # xml_bad_expatreader
    "B316",  # xml_bad_expatbuilder
    "B317",  # xml_bad_sax
    "B318",  # xml_bad_minidom
    "B319",  # xml_bad_pulldom
    "B320",  # xml_bad_etree
    "B321",  # ftplib usage
    "B323",  # unverified_context
    "B324",  # hashlib usage
    "B501",  # request_with_no_cert_validation
    "B502",  # ssl_with_bad_version
    "B503",  # ssl_with_bad_defaults
    "B504",  # ssl_with_no_version
    "B505",  # weak_cryptographic_key
    "B506",  # yaml_load
    "B507",  # ssh_no_host_key_verification
    "B601",  # paramiko_calls
    "B602",  # subprocess_popen_with_shell_equals_true
    "B603",  # subprocess_without_shell_equals_true
    "B604",  # any_other_function_with_shell_equals_true
    "B605",  # start_process_with_a_shell
    "B606",  # start_process_with_no_shell
    "B607",  # start_process_with_partial_path
    "B608",  # hardcoded_sql_expressions
    "B609",  # linux_commands_wildcard_injection
    "B610",  # django_extra_used
    "B611",  # django_rawsql_used
    "B701",  # jinja2_autoescape_false
    "B702",  # use_of_mako_templates
    "B703",  # django_mark_safe
]

# Patterns to detect in code
[tool.bandit.assert_used]
skips = ["*_test.py", "test_*.py"]
